changed IDs from int to long + added related tracks test + moved all models to separate files

+++ SCPlayerTests/DummyDatabase.cs
@@ -21,32 +21,32 @@ namespace SCPlayerTests
         {
         }
 
-        public Task<Track> GetTrackById(int id, CancellationToken token)
+        public Task<Track> GetTrackById(long id, CancellationToken token)
         {
             return Task.FromResult(new Track() { Id = id });
         }
 
-        public Task<User> GetUserById(int id, CancellationToken token)
+        public Task<User> GetUserById(long id, CancellationToken token)
         {
             return Task.FromResult(new User() { Id = id });
         }
 
-        public Task<Playlist> GetPlaylistById(int id, CancellationToken token)
+        public Task<Playlist> GetPlaylistById(long id, CancellationToken token)
         {
             return Task.FromResult(new Playlist() { Id = id });
         }
 
-        public Task<IEnumerable<Track>> GetTracksById(IEnumerable<int> ids, CancellationToken token)
+        public Task<IEnumerable<Track>> GetTracksById(IEnumerable<long> ids, CancellationToken token)
         {
             return Task.FromResult(Enumerable.Empty<Track>());
         }
 
-        public Task<IEnumerable<User>> GetUsersById(IEnumerable<int> ids, CancellationToken token)
+        public Task<IEnumerable<User>> GetUsersById(IEnumerable<long> ids, CancellationToken token)
         {
             return Task.FromResult(Enumerable.Empty<User>());
         }
 
-        public Task<IEnumerable<Playlist>> GetPlaylistsById(IEnumerable<int> ids, CancellationToken token)
+        public Task<IEnumerable<Playlist>> GetPlaylistsById(IEnumerable<long> ids, CancellationToken token)
         {
             return Task.FromResult(Enumerable.Empty<Playlist>());
         }
