+++ SoundMist/Models/BlockedEntry.cs
@@ -1,4 +1,4 @@
 ﻿namespace SoundMist.Models
 {
-    public record BlockedEntry(int Id, string Title);
+    public record BlockedEntry(long Id, string Title);
 }
